pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diseno-multiagente"
version = "0.1.0"
description = "Multi-agent graph architecture: LangGraph + CrewAI + Pydantic + LangSmith"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "pydantic>=2.0",
    "langgraph>=0.2.0",
    "crewai>=0.80.0",
    "langsmith>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
